Implement a parallel blocked cholesky algorithm
This will not be the final commit that includes changes to the cholesky preconditioner.
In particular the triangular solve algorithm that is used when applying the preconditioner still needs to be optimized to run in parallel using SYCL. There were problems with stability but this will be fixed in the future.